He creado paquetes para cada ejercicio y un main para cada uno, he borroado el main principal y solo queda la clase time
