+++ Tests/Java/es/uah/matcomp/mp/e1/ejerciciosclases/Practica1/BallTest.java
@@ -1,5 +1,6 @@
 package es.uah.matcomp.mp.e1.ejerciciosclases.Practica1;
 
+import es.uah.matcomp.mp.e1.ejerciciosclases.Practica1.Ball.Ball;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
