+++ src/es/uah/matcomp/mp/e1/ejerciciosclases/Practica1/Account/Account.java
@@ -1,4 +1,4 @@
-package es.uah.matcomp.mp.e1.ejerciciosclases.Practica1;
+package es.uah.matcomp.mp.e1.ejerciciosclases.Practica1.Account;
 
 public class Account {
