+++ src/es/uah/matcomp/mp/e1/ejerciciosclases/Practica1/Employee/Employee.java
@@ -1,4 +1,4 @@
-package es.uah.matcomp.mp.e1.ejerciciosclases.Practica1;
+package es.uah.matcomp.mp.e1.ejerciciosclases.Practica1.Employee;
 
 public class Employee {
     private int id;
