+++ src/es/uah/matcomp/mp/e1/ejerciciosclases/Practica1/InvoiceItem/InvoiceItem.java
@@ -1,4 +1,4 @@
-package es.uah.matcomp.mp.e1.ejerciciosclases.Practica1;
+package es.uah.matcomp.mp.e1.ejerciciosclases.Practica1.InvoiceItem;
 
 public class InvoiceItem {
     private String id;
